requirements.txt
python-dotenv==0.19.1
openai==0.27.0
fastapi==0.68.0
uvicorn==0.15.0